[stable-only] reno: fix indentation
This missing empty line causes errors time to time. The simplest is to
fix it on every branch so that this won't come up again and again in
the future.

The error is usually something like this:

<reno.sphinxext <some_branch>>:32:Unexpected indentation.

And can be identified with adding extra verbosity to releasenotes job,
via -vvv argument in tox.ini to the sphinx-build command.

Change-Id: I17da28b3ed181531e32a3c2ce3ef445af1d5c978
(cherry picked from commit 254a03a0f2064bdb4addf7727320e33c471ddcc5)